Passes some of the callback tests.
There's no facility for removing them at this point, so all the ones that test
for that fail. The one I don't understand yet why it's failing is
"callbacks_only_fire_on_change", but I just need to look at the test harder.

I suspect that I don't need to register downstream clients with computecells,
and I get exactly-once behavior with callbacks on set_value() due to the caching
of the previous values in the compute cell.